+++ CHANGES.rst
@@ -5,8 +5,9 @@ Changelog
 ------------------
 
 - added convenience methods `no_class` (to unset class) and `has_class` (class set?) to `Instances` class
-- switched to using method objects to speed up methods `classify_instance`/`distribution_for_instance` in `Classifier` class
-- switched to using method objects to speed up methods `cluster_instance`/`distribution_for_instance` in `Clusterer` class
+- switched to using faster method objects for methods `classify_instance`/`distribution_for_instance` in `Classifier` class
+- switched to using faster method objects for methods `cluster_instance`/`distribution_for_instance` in `Clusterer` class
+- switched to using faster method objects for methods `class_index`, `is_missing`, `get/set_value`, `get/set_string_value`, `weight` in `Instance` class
 - ...